Changed datatype to prevent overflows

+++ absp/include/ADQ14_FWPD.hpp
@@ -115,7 +115,7 @@ public:
     std::vector<unsigned int> added_headers;
     std::vector<unsigned int> status_headers;
     std::vector<std::vector<int16_t>> incomplete_records;
-    std::vector<unsigned int> remaining_samples;
+    std::vector<int64_t> remaining_samples;
 
     int64_t transfer_buffer_size;
     unsigned int transfer_buffers_number;
